# Avoid enumerator allocations inside KeyedCollection<TKey, TItem>