@jasonwilliams200OK 2016-01-20
:+1:

Apologies for going slightly off-topic @justinvp; (since you are touching the related code) do you think if it would provide any performance advantage to have a `Dictionary<TKey,TItem>.Add(KeyValuePair<TKey,TItem> pair)` overload, over the existing `Dictionary<TKey,TItem>.Add(TKey key,TItem item)`? When using LINQ or loops, sometimes we end up having `KeyValuePair`s in a lambda, which needs to be inserted in the Dictionary. The current approach is to access the key and value properties individually: `Add(pair.Key, pair.Value)`.