@justinvp 2016-01-20
@jasonwilliams200OK I don't think there'd be a performance advantage. Internally, an `Add` overload that took a `KeyValuePair<TKey, TValue>` would have to call the individual `Key` and `Value` properties to get those values, which is what `Dictionary<TKey, TValue>`'s explicit [implementation of `ICollection<T>.Add`](https://github.com/dotnet/coreclr/blob/d6d14eecdcf2a9cdf0c3a48e48f523f35613749b/src/mscorlib/src/System/Collections/Generic/Dictionary.cs#L208-L210) does. It may even be a perf disadvantage as the `KeyValuePair<TKey, TValue>` would have to be copied each call to `Add` instead of the key and value passed directly.